+++ abis/index.ts
@@ -1,0 +1,19 @@
+import {default as ResourcesAbi} from "./Resources.json"
+import {default as MarketplaceAbi} from "./Marketplace.json"
+import {default as ERC20Abi} from "./ERC20.json"
+import {default as RatingSystemAbi} from "./RatingSystem.json"
+import {default as DisputesAbi} from "./Disputes.json"
+import {default as MarketplaceHelperAbi} from "./MarketplaceHelper.json"
+import {default as MarketplaceViewerAbi} from "./MarketplaceViewer.json"
+import {default as StakingAbi} from "./Staking.json"
+
+export default {
+    ResourcesAbi,
+    MarketplaceAbi,
+    ERC20Abi,
+    RatingSystemAbi,
+    DisputesAbi,
+    MarketplaceHelperAbi,
+    MarketplaceViewerAbi,
+    StakingAbi
+}