+++ src/models/erc20.ts
@@ -1,7 +1,8 @@
 import {Sdk, SdkConfig} from "../config/sdk";
-
-import {abi as ERC20ABI} from "../../abis/ERC20.json";
+import abi from "../../abis/ERC20.json";
 import {Address} from "viem";
+
+const ERC20ABI: typeof abi = abi
 
 export class ERC20 {
   private config: SdkConfig
@@ -13,7 +14,7 @@ export class ERC20 {
     try {
       return await this.config.publicClient.readContract({
         address: address,
-        abi: ERC20ABI,
+        abi: ERC20ABI.abi,
         functionName: functionName,
         args: args
       })
@@ -26,7 +27,7 @@ export class ERC20 {
     try {
       const { request } = await this.config.publicClient.simulateContract({
         address: address,
-        abi: ERC20ABI,
+        abi: ERC20ABI.abi,
         functionName: functionName,
         args: args,
         account: this.config.walletClient.account!.address,