+++ src/models/eventsHandler.ts
@@ -1,11 +1,15 @@
-import * as Addresses from "../../contractAddresses.json";
+import Addresses from "../../contractAddresses.json";
 
-import {abi as ResourcesABI} from "../../abis/Resources.json";
-import {abi as MarketplaceABI} from "../../abis/Marketplace.json";
-import {abi as MarketplaceViewerABI} from "../../abis/MarketplaceViewer.json";
-
+import Resources from "../../abis/Resources.json";
+import Marketplace from "../../abis/MarketplaceViewer.json";
+import MarketplaceViewer from "../../abis/MarketplaceViewer.json";
 import {Sdk, SdkConfig} from "../config/sdk";
 import {Address} from "viem";
+
+const ContractAddresses: typeof Addresses = Addresses
+const ResourcesABI: typeof Resources = Resources
+const MarketplaceABI: typeof Marketplace = Marketplace
+const MarketplaceViewerABI: typeof MarketplaceViewer = MarketplaceViewer
 
 export class EventsHandler {
   private config: SdkConfig
@@ -28,7 +32,7 @@ export class EventsHandler {
     toBlock: bigint,
   }) {
     if (
-      Addresses[contractName][this.config.publicClient.chain!.id] === undefined
+      ContractAddresses[contractName][this.config.publicClient.chain!.id] === undefined
     ) {
       throw new Error(
         contractName +
@@ -37,8 +41,8 @@ export class EventsHandler {
       )
     }
     return await this.config.publicClient.getContractEvents({
-      address: Addresses[contractName][this.config.publicClient.chain!.id] as Address,
-      abi: contractAbi,
+      address: ContractAddresses[contractName][this.config.publicClient.chain!.id] as Address,
+      abi: contractAbi.abi,
       eventName: eventName,
       fromBlock: fromBlock,
       toBlock: toBlock,
@@ -53,7 +57,7 @@ export class EventsHandler {
     onError,
   }) {
     if (
-      Addresses[contractName][this.config.publicClient.chain!.id] === undefined
+      ContractAddresses[contractName][this.config.publicClient.chain!.id] === undefined
     ) {
       throw new Error(
         contractName +
@@ -62,8 +66,8 @@ export class EventsHandler {
       )
     }
     this.config.publicClient.watchContractEvent({
-      address: Addresses[contractName][this.config.publicClient.chain!.id],
-      abi: contractAbi,
+      address: ContractAddresses[contractName][this.config.publicClient.chain!.id],
+      abi: contractAbi.abi,
       eventName: eventName,
       onLogs: (logs) => callback(logs),
       onError: (error) => onError(error),