+++ src/models/ratingSystem.ts
@@ -1,6 +1,8 @@
 import {Sdk, SdkConfig} from "../config/sdk";
 import {Address} from "viem";
-import {abi as RatingSystemABI} from "../../abis/RatingSystem.json"
+import abi from "../../abis/RatingSystem.json"
+
+const RatingSystemABI: typeof abi = abi
 
 export class RatingSystem {
     private config: SdkConfig
@@ -12,7 +14,7 @@ export class RatingSystem {
         try {
             return await this.config.publicClient.readContract({
                 address: address,
-                abi: RatingSystemABI,
+                abi: RatingSystemABI.abi,
                 functionName: functionName,
                 args: args
             })
@@ -25,7 +27,7 @@ export class RatingSystem {
         try {
             const { request } = await this.config.publicClient.simulateContract({
                 address: address,
-                abi: RatingSystemABI,
+                abi: RatingSystemABI.abi,
                 functionName: functionName,
                 args: args,
                 account: this.config.walletClient.account!.address,